move discovery into core service
