+++ src/main/validation.js
@@ -35,12 +35,6 @@ export const PersistedCoordinateFormatSchema = v.union([
 	v.literal('utm'),
 ])
 
-export const DiscoveryInitMessageSchema = v.object({
-	type: v.literal('core:discovery-init'),
-	name: v.string(),
-	port: v.number(),
-})
-
 export const ServiceErrorMessageSchema = v.object({
 	type: v.literal('error'),
 	error: v.instance(Error),
